requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.1
Django==1.11.11
djangorestframework==3.7.7
enum34==1.1.6
idna==2.6
ipaddress==1.0.19
ordereddict==1.1
pathlib==1.0.1
pipenv==11.8.3
pycparser==2.18
pyOpenSSL==17.5.0
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
